Adding menu click sounds
Close #2
Not perfect as some sounds don't play due to the way the project is laid
out.  Need to be smarter next time and make sure the audio is not so
tightly coupled to the object so it can play even when object is removed
from the tree, or when we are restarting a scene.